moved changes to resetPassword
