+++ pages/settings.js
@@ -1,13 +1,8 @@
 import Layout from '../components/Layout';
 import withAuth from '../lib/withAuth';
-import Settings from '../components/settings/Settings';
 
 function Page() {
-  return (
-    <Layout pageName="Settings">
-      <Settings />
-    </Layout>
-  );
+  return <Layout pageName="Settings" />;
 }
 
 export default withAuth(Page);
